requirements.txt
discord.py
pymysql
python-dotenv
pytest
